refactor: move SpringPipeline into distinct mod
